backend/requirements.txt
requests
openai
azure-storage-blob
azure-search-documents

python-dotenv